+++ src/application/Gambler.java
@@ -13,7 +13,6 @@ public class Gambler {
 	 */
 	private int suspicion, chips, ssGameIndex, playerIndex;
 	private String name;
-	private Hand hand;
 	//public static int guid;	
 	//private int uid;
 	public static boolean dead;	// game state, if(dead) SSEnds();
@@ -36,10 +35,10 @@ public class Gambler {
 	/***
 	 * + Gambler(int)
 	 */
-	public Gambler(int chips)
+	public Gambler(String name)
 	{
-		this.chips = chips;
-		this.name = null;
+		this.chips = 10000;
+		this.name = name;
 		this.suspicion = 0;
 		this.ssGameIndex = -1;
 		this.playerIndex = -1;